Fixed IDE errors
