+++ RTClib.h
@@ -23,8 +23,7 @@
 #define _RTCLIB_H_
 
 #include <Arduino.h>
-#include "utility/RTC.h"
-class TimeSpan;
+#include "utility/RTC/RTC.h"
 
 /** Registers */
 #define PCF8523_ADDRESS 0x68       ///< I2C address for PCF8523
@@ -58,210 +57,17 @@ class TimeSpan;
   0x11 ///< Temperature register (high byte - low byte is at 0x12), 10-bit
        ///< temperature value
 
-/** Constants */
-#define SECONDS_PER_DAY 86400L ///< 60 * 60 * 24
-#define SECONDS_FROM_1970_TO_2000                                              \
-  946684800 ///< Unixtime for 2000-01-01 00:00:00, useful for initialization
-
-const DateTime COMPILE_DT = DateTime(__DATE__, __TIME__);
-
-bool isLeapYear(uint16_t year);
-uint8_t getDaysInMonth(uint16_t year, uint8_t month);
-
-/**************************************************************************/
-/*!
-    @brief  Simple general-purpose date/time class (no TZ / DST / leap
-            seconds).
-
-    This class stores date and time information in a broken-down form, as a
-    tuple (year, month, day, hour, minute, second). The day of the week is
-    not stored, but computed on request. The class has no notion of time
-    zones, daylight saving time, or
-    [leap seconds](http://en.wikipedia.org/wiki/Leap_second): time is stored
-    in whatever time zone the user chooses to use.
-
-    The class supports dates in the range from 1 Jan 2000 to 31 Dec 2099
-    inclusive.
-*/
-/**************************************************************************/
-class DateTime {
-public:
-  DateTime(uint32_t t = SECONDS_FROM_1970_TO_2000);
-  DateTime(uint16_t year, uint8_t month, uint8_t day, uint8_t hour = 0,
-           uint8_t min = 0, uint8_t sec = 0);
-  DateTime(const DateTime &copy);
-  DateTime(const char *date, const char *time);
-  DateTime(const __FlashStringHelper *date, const __FlashStringHelper *time);
-  DateTime(const char *iso8601date);
-  bool isValid() const;
-  bool fixDateTime();
-  char *toString(char *buffer);
-
-  /*!
-      @brief  Return the year.
-      @return Year (range: 2000--2099).
-  */
-  uint16_t year() const { return 2000U + yOff; }
-  /*!
-      @brief  Return the month.
-      @return Month number (1--12).
-  */
-  uint8_t month() const { return m; }
-  /*!
-      @brief  Return the day of the month.
-      @return Day of the month (1--31).
-  */
-  uint8_t day() const { return d; }
-  /*!
-      @brief  Return the hour
-      @return Hour (0--23).
-  */
-  uint8_t hour() const { return hh; }
-
-  uint8_t twelveHour() const;
-  /*!
-      @brief  Return whether the time is PM.
-      @return 0 if the time is AM, 1 if it's PM.
-  */
-  uint8_t isPM() const { return hh >= 12; }
-  /*!
-      @brief  Return the minute.
-      @return Minute (0--59).
-  */
-  uint8_t minute() const { return mm; }
-  /*!
-      @brief  Return the second.
-      @return Second (0--59).
-  */
-  uint8_t second() const { return ss; }
-
-  uint8_t dayOfTheWeek() const;
-
-  /* 32-bit times as seconds since 2000-01-01. */
-  uint32_t secondstime() const;
-
-  /* 32-bit times as seconds since 1970-01-01. */
-  uint32_t unixtime(void) const;
-
-  /*!
-      Format of the ISO 8601 timestamp generated by `timestamp()`. Each
-      option corresponds to a `toString()` format as follows:
-  */
-  enum timestampOpt {
-    TIMESTAMP_FULL, //!< `YYYY-MM-DDThh:mm:ss`
-    TIMESTAMP_TIME, //!< `hh:mm:ss`
-    TIMESTAMP_DATE  //!< `YYYY-MM-DD`
-  };
-  String timestamp(timestampOpt opt = TIMESTAMP_FULL);
-
-  DateTime operator+(const TimeSpan &span);
-  DateTime operator-(const TimeSpan &span);
-  TimeSpan operator-(const DateTime &right);
-  bool operator<(const DateTime &right) const;
-
-  /*!
-      @brief  Test if one DateTime is greater (later) than another.
-      @warning if one or both DateTime objects are invalid, returned value is
-        meaningless
-      @see use `isValid()` method to check if DateTime object is valid
-      @param right DateTime object to compare
-      @return True if the left DateTime is later than the right one,
-        false otherwise
-  */
-  bool operator>(const DateTime &right) const { return right < *this; }
-
-  /*!
-      @brief  Test if one DateTime is less (earlier) than or equal to another
-      @warning if one or both DateTime objects are invalid, returned value is
-        meaningless
-      @see use `isValid()` method to check if DateTime object is valid
-      @param right DateTime object to compare
-      @return True if the left DateTime is earlier than or equal to the
-        right one, false otherwise
-  */
-  bool operator<=(const DateTime &right) const { return !(*this > right); }
-
-  /*!
-      @brief  Test if one DateTime is greater (later) than or equal to another
-      @warning if one or both DateTime objects are invalid, returned value is
-        meaningless
-      @see use `isValid()` method to check if DateTime object is valid
-      @param right DateTime object to compare
-      @return True if the left DateTime is later than or equal to the right
-        one, false otherwise
-  */
-  bool operator>=(const DateTime &right) const { return !(*this < right); }
-  bool operator==(const DateTime &right) const;
-
-  /*!
-      @brief  Test if two DateTime objects are not equal.
-      @warning if one or both DateTime objects are invalid, returned value is
-        meaningless
-      @see use `isValid()` method to check if DateTime object is valid
-      @param right DateTime object to compare
-      @return True if the two objects are not equal, false if they are
-  */
-  bool operator!=(const DateTime &right) const { return !(*this == right); }
-
-protected:
-  uint8_t yOff; ///< Year offset from 2000
-  uint8_t m;    ///< Month 1-12
-  uint8_t d;    ///< Day 1-31
-  uint8_t hh;   ///< Hours 0-23
-  uint8_t mm;   ///< Minutes 0-59
-  uint8_t ss;   ///< Seconds 0-59
-};
-
-/**************************************************************************/
-/*!
-    @brief  Timespan which can represent changes in time with seconds accuracy.
-*/
-/**************************************************************************/
-class TimeSpan {
-public:
-  TimeSpan(int32_t seconds = 0);
-  TimeSpan(int16_t days, int8_t hours, int8_t minutes, int8_t seconds);
-  TimeSpan(const TimeSpan &copy);
-
-  /*!
-      @brief  Number of days in the TimeSpan
-              e.g. 4
-      @return int16_t days
-  */
-  int16_t days() const { return _seconds / 86400L; }
-  /*!
-      @brief  Number of hours in the TimeSpan
-              This is not the total hours, it includes the days
-              e.g. 4 days, 3 hours - NOT 99 hours
-      @return int8_t hours
-  */
-  int8_t hours() const { return _seconds / 3600 % 24; }
-  /*!
-      @brief  Number of minutes in the TimeSpan
-              This is not the total minutes, it includes days/hours
-              e.g. 4 days, 3 hours, 27 minutes
-      @return int8_t minutes
-  */
-  int8_t minutes() const { return _seconds / 60 % 60; }
-  /*!
-      @brief  Number of seconds in the TimeSpan
-              This is not the total seconds, it includes the days/hours/minutes
-              e.g. 4 days, 3 hours, 27 minutes, 7 seconds
-      @return int8_t seconds
-  */
-  int8_t seconds() const { return _seconds % 60; }
-  /*!
-      @brief  Total number of seconds in the TimeSpan, e.g. 358027
-      @return int32_t seconds
-  */
-  int32_t totalseconds() const { return _seconds; }
-
-  TimeSpan operator+(const TimeSpan &right);
-  TimeSpan operator-(const TimeSpan &right);
-
-protected:
-  int32_t _seconds; ///< Actual TimeSpan value is stored as seconds
-};
+#if (ARDUINO >= 100)
+#include <Arduino.h> // capital A so it is error prone on case-sensitive filesystems
+// Macro to deal with the difference in I2C write functions from old and new
+// Arduino versions.
+#define _I2C_WRITE write ///< Modern I2C write
+#define _I2C_READ read   ///< Modern I2C read
+#else
+#include <WProgram.h>
+#define _I2C_WRITE send   ///< Legacy I2C write
+#define _I2C_READ receive ///< legacy I2C read
+#endif
 
 /** DS1307 SQW pin mode settings */
 enum Ds1307SqwPinMode {
@@ -280,14 +86,14 @@ enum Ds1307SqwPinMode {
 /**************************************************************************/
 class RTC_DS1307 : RTC {
 public:
-  boolean begin(const DateTime &dt = COMPILE_DT);
-  static void adjust(const DateTime &dt);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
+  void adjust(const DateTime &dt);
   void adjustDrift(int drift);
   bool isrunning(void);
   bool lostPower(void);
-  static DateTime now();
-  static Ds1307SqwPinMode readSqwPinMode();
-  static void writeSqwPinMode(Ds1307SqwPinMode mode);
+  DateTime now();
+  Ds1307SqwPinMode readSqwPinMode();
+  void writeSqwPinMode(Ds1307SqwPinMode mode);
   uint8_t readnvram(uint8_t address);
   void readnvram(uint8_t *buf, uint8_t size, uint8_t address);
   void writenvram(uint8_t address, uint8_t data);
@@ -334,14 +140,14 @@ enum Ds3231Alarm2Mode {
 /**************************************************************************/
 class RTC_DS3231 : RTC {
 public:
-  boolean begin(const DateTime &dt = COMPILE_DT);
-  static void adjust(const DateTime &dt);
-  static void adjustDrift(const int drift);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
+  void adjust(const DateTime &dt);
+  void adjustDrift(const int drift);
   bool isrunning(void);
   bool lostPower(void);
-  static DateTime now();
-  static Ds3231SqwPinMode readSqwPinMode();
-  static void writeSqwPinMode(Ds3231SqwPinMode mode);
+  DateTime now();
+  Ds3231SqwPinMode readSqwPinMode();
+  void writeSqwPinMode(Ds3231SqwPinMode mode);
   bool setAlarm1(const DateTime &dt, Ds3231Alarm1Mode alarm_mode);
   bool setAlarm2(const DateTime &dt, Ds3231Alarm2Mode alarm_mode);
   void disableAlarm(uint8_t alarm_num);
@@ -401,12 +207,12 @@ enum Pcf8523OffsetMode {
 /**************************************************************************/
 class RTC_PCF8523 : RTC {
 public:
-  boolean begin(const DateTime &dt = COMPILE_DT);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
   void adjust(const DateTime &dt);
   boolean isrunning(void);
   boolean lostPower(void);
   boolean initialized(void);
-  static DateTime now();
+  DateTime now();
   void start(void);
   void stop(void);
   boolean isrunning();
@@ -439,14 +245,14 @@ enum Pcf8563SqwPinMode {
 
 class RTC_PCF8563 : RTC {
 public:
-  boolean begin(const DateTime &dt = COMPILE_DT);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
   boolean isrunning(void);
   boolean lostPower(void);
   void adjust(const DateTime &dt);
-  static DateTime now();
+  DateTime now();
   void start(void);
   void stop(void);
-  uint8_t isrunning();
+  boolean isrunning();
   Pcf8563SqwPinMode readSqwPinMode();
   void writeSqwPinMode(Pcf8563SqwPinMode mode);
 };
@@ -459,7 +265,7 @@ public:
 /**************************************************************************/
 class RTC_Millis : RTC {
 public:
-  static boolean begin(const DateTime &dt = COMPILE_DT);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
   /*!
     @brief  Simulate if the RTC is running
     @return true
@@ -470,9 +276,9 @@ public:
     @return false
   */
   boolean lostPower(void) { return false; }
-  static void adjust(const DateTime &dt);
-  static void adjustDrift(const int drift);
-  static DateTime now();
+  void adjust(const DateTime &dt);
+  void adjustDrift(const int drift);
+  DateTime now();
 
 protected:
   static uint32_t millisPerSecond; ///< Number of milliseconds reported by
@@ -498,7 +304,7 @@ public:
       @brief  Start the RTC
       @param dt DateTime object with the date/time to set
   */
-  static boolean begin(const DateTime &dt = COMPILE_DT);
+  boolean begin(const DateTime &dt = DateTime::COMPILE_DT);
   /*!
     @brief  Simulate if the RTC is running
     @return true
@@ -509,9 +315,9 @@ public:
     @return false
   */
   boolean lostPower(void) { return false; }
-  static void adjust(const DateTime &dt);
-  static void adjustDrift(int ppm);
-  static DateTime now();
+  void adjust(const DateTime &dt);
+  void adjustDrift(int ppm);
+  DateTime now();
 
 protected:
   static uint32_t microsPerSecond; ///< Number of microseconds reported by
