+++ utility/DateTime/DateTime.h
@@ -1,0 +1,163 @@
+#ifndef _DATETIME_H_
+#define _DATETIME_H_
+
+#include <Arduino.h>
+#include "../TimeSpan/TimeSpan.h"
+
+/** Constants */
+#define SECONDS_PER_DAY 86400L ///< 60 * 60 * 24
+#define SECONDS_FROM_1970_TO_2000                                              \
+  946684800 ///< Unixtime for 2000-01-01 00:00:00, useful for initialization
+
+uint8_t getDaysInMonth(uint16_t year, uint8_t month);
+
+/**************************************************************************/
+/*!
+    @brief  Simple general-purpose date/time class (no TZ / DST / leap
+            seconds).
+
+    This class stores date and time information in a broken-down form, as a
+    tuple (year, month, day, hour, minute, second). The day of the week is
+    not stored, but computed on request. The class has no notion of time
+    zones, daylight saving time, or
+    [leap seconds](http://en.wikipedia.org/wiki/Leap_second): time is stored
+    in whatever time zone the user chooses to use.
+
+    The class supports dates in the range from 1 Jan 2000 to 31 Dec 2099
+    inclusive.
+*/
+/**************************************************************************/
+class DateTime {
+public:
+  /*!
+    @brief The date/time at compile time
+  */
+  const static DateTime COMPILE_DT;
+
+  DateTime(uint32_t t = SECONDS_FROM_1970_TO_2000);
+  DateTime(uint16_t year, uint8_t month, uint8_t day, uint8_t hour = 0,
+           uint8_t min = 0, uint8_t sec = 0);
+  DateTime(const DateTime &copy);
+  DateTime(const char *date, const char *time);
+  DateTime(const __FlashStringHelper *date, const __FlashStringHelper *time);
+  DateTime(const char *iso8601date);
+  bool isValid() const;
+  bool fixDateTime();
+  char *toString(char *buffer);
+
+  /*!
+      @brief  Return the year.
+      @return Year (range: 2000--2099).
+  */
+  uint16_t year() const { return 2000U + yOff; }
+  /*!
+      @brief  Return the month.
+      @return Month number (1--12).
+  */
+  uint8_t month() const { return m; }
+  /*!
+      @brief  Return the day of the month.
+      @return Day of the month (1--31).
+  */
+  uint8_t day() const { return d; }
+  /*!
+      @brief  Return the hour
+      @return Hour (0--23).
+  */
+  uint8_t hour() const { return hh; }
+
+  uint8_t twelveHour() const;
+  /*!
+      @brief  Return whether the time is PM.
+      @return 0 if the time is AM, 1 if it's PM.
+  */
+  uint8_t isPM() const { return hh >= 12; }
+  /*!
+      @brief  Return the minute.
+      @return Minute (0--59).
+  */
+  uint8_t minute() const { return mm; }
+  /*!
+      @brief  Return the second.
+      @return Second (0--59).
+  */
+  uint8_t second() const { return ss; }
+
+  uint8_t dayOfTheWeek() const;
+
+  /* 32-bit times as seconds since 2000-01-01. */
+  uint32_t secondstime() const;
+
+  /* 32-bit times as seconds since 1970-01-01. */
+  uint32_t unixtime(void) const;
+
+  /*!
+      Format of the ISO 8601 timestamp generated by `timestamp()`. Each
+      option corresponds to a `toString()` format as follows:
+  */
+  enum timestampOpt {
+    TIMESTAMP_FULL, //!< `YYYY-MM-DDThh:mm:ss`
+    TIMESTAMP_TIME, //!< `hh:mm:ss`
+    TIMESTAMP_DATE  //!< `YYYY-MM-DD`
+  };
+  String timestamp(timestampOpt opt = TIMESTAMP_FULL);
+
+  DateTime operator+(const TimeSpan &span);
+  DateTime operator-(const TimeSpan &span);
+  TimeSpan operator-(const DateTime &right);
+  bool operator<(const DateTime &right) const;
+
+  /*!
+      @brief  Test if one DateTime is greater (later) than another.
+      @warning if one or both DateTime objects are invalid, returned value is
+        meaningless
+      @see use `isValid()` method to check if DateTime object is valid
+      @param right DateTime object to compare
+      @return True if the left DateTime is later than the right one,
+        false otherwise
+  */
+  bool operator>(const DateTime &right) const { return right < *this; }
+
+  /*!
+      @brief  Test if one DateTime is less (earlier) than or equal to another
+      @warning if one or both DateTime objects are invalid, returned value is
+        meaningless
+      @see use `isValid()` method to check if DateTime object is valid
+      @param right DateTime object to compare
+      @return True if the left DateTime is earlier than or equal to the
+        right one, false otherwise
+  */
+  bool operator<=(const DateTime &right) const { return !(*this > right); }
+
+  /*!
+      @brief  Test if one DateTime is greater (later) than or equal to another
+      @warning if one or both DateTime objects are invalid, returned value is
+        meaningless
+      @see use `isValid()` method to check if DateTime object is valid
+      @param right DateTime object to compare
+      @return True if the left DateTime is later than or equal to the right
+        one, false otherwise
+  */
+  bool operator>=(const DateTime &right) const { return !(*this < right); }
+  bool operator==(const DateTime &right) const;
+
+  /*!
+      @brief  Test if two DateTime objects are not equal.
+      @warning if one or both DateTime objects are invalid, returned value is
+        meaningless
+      @see use `isValid()` method to check if DateTime object is valid
+      @param right DateTime object to compare
+      @return True if the two objects are not equal, false if they are
+  */
+  bool operator!=(const DateTime &right) const { return !(*this == right); }
+
+protected:
+  uint8_t yOff; ///< Year offset from 2000
+  uint8_t m;    ///< Month 1-12
+  uint8_t d;    ///< Day 1-31
+  uint8_t hh;   ///< Hours 0-23
+  uint8_t mm;   ///< Minutes 0-59
+  uint8_t ss;   ///< Seconds 0-59
+};
+
+#endif
