+++ utility/TimeSpan/TimeSpan.cpp
@@ -1,0 +1,55 @@
+#include "TimeSpan.h"
+
+/**************************************************************************/
+/*!
+    @brief  Create a new TimeSpan object in seconds
+    @param seconds Number of seconds
+*/
+/**************************************************************************/
+TimeSpan::TimeSpan(int32_t seconds) : _seconds(seconds) {}
+
+/**************************************************************************/
+/*!
+    @brief  Create a new TimeSpan object using a number of
+   days/hours/minutes/seconds e.g. Make a TimeSpan of 3 hours and 45 minutes:
+   new TimeSpan(0, 3, 45, 0);
+    @param days Number of days
+    @param hours Number of hours
+    @param minutes Number of minutes
+    @param seconds Number of seconds
+*/
+/**************************************************************************/
+TimeSpan::TimeSpan(int16_t days, int8_t hours, int8_t minutes, int8_t seconds)
+    : _seconds((int32_t)days * 86400L + (int32_t)hours * 3600 +
+               (int32_t)minutes * 60 + seconds) {}
+
+/**************************************************************************/
+/*!
+    @brief  Copy constructor, make a new TimeSpan using an existing one
+    @param copy The TimeSpan to copy
+*/
+/**************************************************************************/
+TimeSpan::TimeSpan(const TimeSpan &copy) : _seconds(copy._seconds) {}
+
+/**************************************************************************/
+/*!
+    @brief  Add two TimeSpans
+    @param right TimeSpan to add
+    @return New TimeSpan object, sum of left and right
+*/
+/**************************************************************************/
+TimeSpan TimeSpan::operator+(const TimeSpan &right) {
+  return TimeSpan(_seconds + right._seconds);
+}
+
+/**************************************************************************/
+/*!
+    @brief  Subtract a TimeSpan
+    @param right TimeSpan to subtract
+    @return New TimeSpan object, right subtracted from left
+*/
+/**************************************************************************/
+TimeSpan TimeSpan::operator-(const TimeSpan &right) {
+  return TimeSpan(_seconds - right._seconds);
+}
+
