+++ utility/TimeSpan/TimeSpan.h
@@ -1,0 +1,55 @@
+#ifndef _TIMESPAN_H_
+#define _TIMESPAN_H_
+
+/**************************************************************************/
+/*!
+    @brief  Timespan which can represent changes in time with seconds accuracy.
+*/
+/**************************************************************************/
+class TimeSpan {
+public:
+  TimeSpan(int32_t seconds = 0);
+  TimeSpan(int16_t days, int8_t hours, int8_t minutes, int8_t seconds);
+  TimeSpan(const TimeSpan &copy);
+
+  /*!
+      @brief  Number of days in the TimeSpan
+              e.g. 4
+      @return int16_t days
+  */
+  int16_t days() const { return _seconds / 86400L; }
+  /*!
+      @brief  Number of hours in the TimeSpan
+              This is not the total hours, it includes the days
+              e.g. 4 days, 3 hours - NOT 99 hours
+      @return int8_t hours
+  */
+  int8_t hours() const { return _seconds / 3600 % 24; }
+  /*!
+      @brief  Number of minutes in the TimeSpan
+              This is not the total minutes, it includes days/hours
+              e.g. 4 days, 3 hours, 27 minutes
+      @return int8_t minutes
+  */
+  int8_t minutes() const { return _seconds / 60 % 60; }
+  /*!
+      @brief  Number of seconds in the TimeSpan
+              This is not the total seconds, it includes the days/hours/minutes
+              e.g. 4 days, 3 hours, 27 minutes, 7 seconds
+      @return int8_t seconds
+  */
+  int8_t seconds() const { return _seconds % 60; }
+  /*!
+      @brief  Total number of seconds in the TimeSpan, e.g. 358027
+      @return int32_t seconds
+  */
+  int32_t totalseconds() const { return _seconds; }
+
+  TimeSpan operator+(const TimeSpan &right);
+  TimeSpan operator-(const TimeSpan &right);
+
+protected:
+  int32_t _seconds; ///< Actual TimeSpan value is stored as seconds
+};
+
+#endif
